Move simulation stuff to its own file
